Move VanishRequestsState to commons/commonMain, add observeNotes to ICacheProvider
- VanishRequestsState + VanishEventItem + ComplianceStatus: moved to commons/commonMain
- Added observeNotes(Filter) to ICacheProvider interface
- LocalCache.observeNotes now overrides the interface method
- Replaced Dispatchers.IO with Dispatchers_IO for multiplatform compat
