+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip62Vanish/VanishRequestsState.kt
@@ -20,114 +20,7 @@
  */
 package com.vitorpamplona.amethyst.model.nip62Vanish
 
-import androidx.compose.runtime.Stable
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
-import com.vitorpamplona.quartz.nip01Core.relay.client.accessories.fetchFirst
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip62RequestToVanish.RequestToVanishEvent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.withContext
-
-@Stable
-data class VanishEventItem(
-    val event: RequestToVanishEvent,
-    val relays: List<NormalizedRelayUrl>,
-    val isAllRelays: Boolean,
-    val complianceResults: MutableStateFlow<Map<NormalizedRelayUrl, ComplianceStatus>> =
-        MutableStateFlow(
-            relays.associateWith { ComplianceStatus.UNTESTED },
-        ),
-)
-
-enum class ComplianceStatus {
-    UNTESTED,
-    TESTING,
-    COMPLIANT,
-    NON_COMPLIANT,
-    ERROR,
-}
-
-class VanishRequestsState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val client: INostrClient,
-    val scope: CoroutineScope,
-) {
-    val noteFlow =
-        cache
-            .observeNotes(
-                Filter(
-                    kinds = listOf(RequestToVanishEvent.KIND),
-                    authors = listOf(signer.pubKey),
-                ),
-            ).stateIn(
-                scope = scope,
-                started = SharingStarted.Eagerly,
-                initialValue = emptyList(),
-            )
-
-    val testableFlow =
-        noteFlow
-            .map { notes ->
-                notes.mapNotNull {
-                    val noteEvent = it.event
-                    if (noteEvent is RequestToVanishEvent) {
-                        VanishEventItem(noteEvent, noteEvent.vanishFromRelays(), noteEvent.vanishFromAllRelays())
-                    } else {
-                        null
-                    }
-                }
-            }.stateIn(
-                scope = scope,
-                started = SharingStarted.WhileSubscribed(10000),
-                initialValue = emptyList(),
-            )
-
-    suspend fun testVanishCompliance(
-        item: VanishEventItem,
-        relay: NormalizedRelayUrl,
-    ) {
-        item.complianceResults.update {
-            it + (relay to ComplianceStatus.TESTING)
-        }
-
-        try {
-            val foundEvent =
-                withContext(Dispatchers.IO) {
-                    client.fetchFirst(
-                        relay = relay,
-                        filter =
-                            Filter(
-                                authors = listOf(item.event.pubKey),
-                                until = item.event.createdAt - 1,
-                                limit = 1,
-                            ),
-                    )
-                }
-
-            item.complianceResults.update {
-                it + (
-                    relay to
-                        if (foundEvent != null) {
-                            ComplianceStatus.NON_COMPLIANT
-                        } else {
-                            ComplianceStatus.COMPLIANT
-                        }
-                )
-            }
-        } catch (_: Exception) {
-            item.complianceResults.update {
-                it + (relay to ComplianceStatus.ERROR)
-            }
-        }
-    }
-}
+// Re-export from commons for backward compatibility
+typealias VanishEventItem = com.vitorpamplona.amethyst.commons.model.nip62Vanish.VanishEventItem
+typealias ComplianceStatus = com.vitorpamplona.amethyst.commons.model.nip62Vanish.ComplianceStatus
+typealias VanishRequestsState = com.vitorpamplona.amethyst.commons.model.nip62Vanish.VanishRequestsState
